refactor(test): introduce utils_uffd.py
Currently, all tests that want to make use of UFFD import UFFD related
functions from test_uffd.py. That's a bit awkward, so move them into
their own utility module.
